[Pkmn Stadium] Protect "House Advantage" against triggering in non-battle modes
Because the "House Advantage" cheevo was meant to work in any battle, it didn't
check the usual "mode" flags to constrain its logic. This ended up being a
mistake, since it could trigger easily by, for example, exiting from a minigame
to the plaza.

Fixed by adding a requirement to the achievement: now it can only be obtained
in a Stadium or Gym Leader Castle battle.