[Pkmn Stadium] Increase "Water at the End of the Tunnel"'s threshold to 6.5s
The cheevo was originally developed using a nearly-correct timer. Using that
timer, the 6s were achievable somewhat easily, which was the intent given the
cheevo is only worth 3 points. However, whilst developing the leaderboards, a
better and more correct timer was found. Using the new timer, the 6s threshold
is way too harsh.

The threshold was increased to 6.5 seconds, which better represents the
intended difficulty.