[Pkmn Stadium] Use PP values for "The Struggle is Real"
The cheevo used to work based on the value of an address that identified the
current move being used. Unfortunately, while the address is static, it differs
between the different cups and gyms.

Attempted to change the cheevo to use the per-pokemon current move address, but
that didn't work, because Struggle's id does not show up there. The id simply
stays at the last "proper" move used by the Pokemon.

The final solution was to use the per-pokemon PP values. If the pokemon has no
PP left on any of its moves, we assume it just used struggle. This not perfect,
since the pokemon could potentially knock itself out with its literal last PP
and trigger the cheevo, but the odds are astronomical.

Two more changes:

- Instead of checking that the hp text was "  1/" in the previous frame, the
cheevo now checks that it wasn't "  0/". This caused a bunch of ALTs to be
generated, but should be much safer against failed cheevo triggers.

- The cheevo can now only be triggered in a Stadium or Gym Leader Castle battle.
This constraints the logic, preventing hypothetical mistriggers when in
non-battle modes.