[pokemon stadium 2] [17507] rewrite achievement logic
The achievement logic was brittle in the sense that it used a very
simple reset condition (a move being used) and then relied on a large
sequence of exception conditions to avoid that reset whenever (1) a move
was used by the opponent, or (2) a move was used by the player that was
part of the moveset of the rental pokemon that make up Red's team.

Using moves to check if the team is correct also had several issues:

* The achievement had to be restricted to rental pokemon, because
  those are the only ones whose moves are fixed and known.

* The team restriction can be bypass by players that know how the
  achievement is coded.

* The achievement's reset condition must be active for the whole
  duration of the cup, which compounds the brittleness issues. There
  are a lot of opportunities for an unforeseen scenario to mess up the
  logic.

----

Rewrote the achievement logic to avoid the highlighted issues by making
it simpler: when the last fight of a stadium cup is won, trigger if
the team is the correct one (using the team's pokemon ids).

The only downside of the new logic is that the team must be in the
correct order, which doesn't seem that bad. Allowing all possible orders
would require either an impossible number of ALT groups, or an
approximate solution that could be abused to use other teams.