[aria of sorrow] [20319] consider souls from copied save game
The issue with the soul-related achievements was that when a save game
is copied, the souls are not carried over in the "normal" memory
addresses. Instead, they are stored in a same-size array after the
normal array. So, the souls layout in memory is:
  - 28 bytes for red souls
  - 28 bytes for red souls from the copied save game
  - 12 bytes for blue souls
  - 1 byte with unknown meaning
  - 12 bytes for blue souls from the copied save game
  - 18 bytes for yellow souls
  - 18 bytes for yellow souls from the copied save game

Fixed by summing both the "normal" & "copied" soul amounts, and using
that as the real amount value for the achievement. This sum is what the
game display in the menu.

Bonus: replaced the hit-count-based save protection with delta-based
save protection.