[aria of sorrow] [20660] handle gold as 32-bit value
The issue was related to how the gold value was being used. The value is
32-bit, but it was being used in two 16-bit conditions instead of a
single 32-bit condition. This meant that it was feasible for a player
to get enough gold past 200k in a single item sale to overflow the
condition that was trying to check for >=200k. It was also feasible for
the player to have a pre-200k amount of gold that caused the
save-protection condition to keep the cheevo from triggering.

The cheevo was also not properly save-protected.

----

Fixed by using the gold address in proper 32-bit conditions. Added save
protection using the same method already used by other achievements of
the set (reset if the past frame's hp value was 0, i.e, reset while not
ingame).