[aria of sorrow] [19975] copy/delete-data-proof in-game check
The existing in-game check was using the player's max hp value, under
the assumption that it could only be non-zero during gameplay. This was
not correct, however: the max hp value is part of the SRAM, which is
loaded into memory when a save game is copied or deleted, both
operations that occur in the main menu, not in-game.

Character stats do not suffer from this problem, because they're
calculated, not stored in SRAM. As such, the in-game check not uses the
LCK stat value (could use any stat, but for good luck...)

The in-game check is part of the battery protection of the game, so most
achievements will have to be updated with this change later on.