[daedalian opus] [22951] fix timer conditions
All time-based achievements would not trigger if the seconds unit was 0.
This happened because there was a condition on that value being non-zero
to ensure the achievements only triggered after the level was completed
(timer is reset to zero before the level, but keeps its value after the
level is beaten).

Fixed by:

1. Replacing the seconds-unit non-zero condition with a condition that
   verifies looks at the whole timer, not just the seconds' units.

2. Updating the time-less-or-equal-than conditions to use the whole
   timer instead of just looking at part of it.