[daedalian opus] [22952] use animation from final cutscene to trigger
For some reason, the stage id address is incremented before entering the
last level, whilst for all other stages, the value is only incremented
upon entering the stage. As a result, the usual trigger of a
level-completed achievement does not work in this case, because it
triggers in the overworld before entering the target stage.

Fixed by adding conditions that detect the final cutscene. Not the
most elegant solution, but it does work. More specifically, we detect
the animation of the fairy in the cutscene. The animation is split in 4
parts, each of which seemed enough, but just for added safety the
achievement logic checks 2 of them.