[looney tunes] [cheat protection] disallow life refill cheat
The cheat can be detected by the amount of hearts increasing whilst the
game is paused.

The protection was included in all achievements except those in that
deal with Taz's stage, because life is irrelevant in that stage.

For beat-the-stage achievements, the protection was included by having
an alt group pause forever if the cheat is used. The pause is reset by
the core group whenever the player leaves the stage.

For beat-boss-damageless achievements, the protection was included by
simply reseting when the cheat is used. This is possible because these
achievements were already scoped. Same goes for the "beat the game w/o
using continues" achievement.