[tiger-heli] [cheat protection] disallow continue cheat
The continue cheat allows the player to continue indefinitely by holding
A+B on the game over screen. This can be detected by the fact that lives
increase from 0 to 2 whilst not transitioning from the title screen to
in-game.

Protection was added to all achievements using a persistent PauseIf (hit
target of 1) that locks the player out of unlocking the achievements if
the cheat is ever used. The PauseIf (and thus the protection) is reset
if the player goes back to the title screen.