[trojan] [cheat protection] disallow continue cheat
The continue cheat allows the player to continue from the furthest
reached stage after a game over by pressing Up + Start in the title
screen.

The cheat is detected by the fact that the up button is pressed during
the transition from the title screen to in-game. A first attempt tried
to use the stage value during the same transition, but that proved
impossible due to the fact that the stage value is updated 1 frame after
the in-game/title-screen flag.

Protection was added to all achievements using a persistent PauseIf (hit
target of 1) that locks the player out of unlocking the achievements if
the cheat is ever used. The PauseIf (and thus the protection) is reset
if the player goes back to the title screen.