[daedalian opus] [23197] use animation from post-35 cutscene to trigger
Achievement was broken by the fix of 22951. The original code was also
broken under certain, though rare, conditions.

The other beat-the-level achievements check that a certain address is
zero to detect the post-level overworld, but for some reason that
address behaves differently between levels 35 and 36. The original
version of the achievement (before the 22951 fix) used the same address,
but looked for value 0x7e. This worked as long as the player waited for
a few seconds before leaving the level after beating it. If that time
was not elapsed, the address would never reach the magic value and the
achievement would not trigger.

Fixed using the same trick used to fix 22952: the end of level 35 is
detected using the animation of the fairy that appears in the post-level
cutscene.