[dino crisis] [23824] reset "Am I Late?" hits when not in target room
The achievement registers a hit when the cutscene starts and triggers
when it ends. The event flag is, however, shared with the alternative
cutscene in which Regina is not "late". There is, thus, some extra
conditions to ensure the achievement does not trigger in this case.

The problem was that the hit registered when the cutscene started was
never reset. In most cases this would have been fine due to the in-room
checks, but due to sheer bad luck the room in which the cutscene takes
place has id 0 (in floor 6). It just so happens that during the final
debriefing, the room id is reset to 0, but the floor id is not. Since
the ending takes place in floor 6, during the debriefing we're in the
correct room according to the floor and room ids. Coupled with the
leftover hit, the achievement would trigger when the in-cutscene flag
was cleared during the fadeout.

Fixed by resetting hits when we're not in the correct room, which was
already done for many other hit-using achievements.