[soul reaver][27302] ensure same instance when testing it with prev
Multiple achievements that operate on instances, namely pool instances,
had checks relying on the state of the instances on the previous frame.
They were not, however, making sure that the pool instance was already
there on the last frame, which could lead to unexpected triggers.

All such achievements now make sure that they target the instances of
interest, but also that those instances were already in place on the
previous frame.