[dino crisis][27406] avoid mistrigger of unlock-operation-wipeout
The achievement would still trigger even if the player had not unlocked
operation wipeout, because the debriefing sequence would still go into
the operation-wipeout-related stage for 2 frames.

Fixed by requiring the debriefing stage to be active for 0.5 seconds.
Should be safe because it doesn't seem possible to skip the stage that
fast.