[pkmn stadium][28745] fix cup cheevos when save-quitting on 7th battle
The memory address being used to identify which cup battle the player is
on gets incremented when the player saves and quits from a cup. This
meant that saving and quitting after the 7th battle would trigger the
respective beat-cup achievement, because the win conditions of the 7th
battle were still active when the battle number got incremented to 8.

Fixed by putting the win-flag condition of beat-cup achievements behind
a delta. This means that the achievements can only trigger on the exact
frame that the battle ends, bypassing the save-and-quit issue.