[sm74ee] fix last 3 frames of blue coin achievements
The blue coin achievements were using the deactivation of the coin block
as the cutoff point, meaning that coins would have to be collected while
the block was active to count towards the achievement. This had a small
issue, however: it turns out the the coin block deactivates 3 frames
before its blue coins, so it was possible for the player to collect the
last coin during one of these 3 frames and not get the achievement to
trigger.

Fixed by considering a blue coin collected if it deactivates on a frame
in which mario's number of collected coins increased by 5 or more. It's
not a perfect solution, but given the context it's extremely unlikely
to be exploitable.