[TR2][33517] avoid corner bug detection on the first frame after load
When loading, Lara's position (and a lot of other data) isn't reset. As
such, on the first post-load frame, Lara's vertical position is suddenly
updated from whatever it was before the load to the save's starting
position. If, in addition, Lara was either running, upwards jumping, or
swimming in that frame, the sudden vertical position update could be
perceived as a corner bug, if it was in the upwards direction.

The obvious fix is to avoid triggering on the first frame after a load.
The usual "is-loading" flag didn't work though, because it deactivates
one frame before the level entities are initialized (including lara's
vertical position). The music track id was used instead: it is
initialized on the same frame as the level's entities, and is always 0
during the load process.