[TR2][33517] avoid corner bug detection when mounting floating vehicles
In Tibetan Foothills, if a snowmobile is parked on the clear ice next
to the hut key, it ends up floating in the air when we progress far
enough for the ice to disappear. When such a floating vehicle was
mounted, the corner bug detection would trigger because: (1) lara
instantly teleports up to the location of the vehicle, and (2) the
action ids while riding a vehicle overlap with the normal ones
(specifically the running action, which has a very low id value).

Fixed by not detecting corner bugs if lara is in a vehicle.

Also turns out the fix in e2158698ec4223dec54332cc56e852146a7758b8 was
not correct because the extra checks added to the corner bug pause lock
need to be inside the `unless(once(...))`s. Fixed that as well.