[TR2][33530] disregard level flip when picking up tibetan mask
"No, You're Locked in Here With Me" stores a hit when the level flips
as a result of pulling the lever in the yeti jail that releases them.
It then trigger when the 4 yetis are inactive, which only happens when
they're killed, assuming we've already pulled the lever.

The issue was that the level also flips when the first tibetan mask is
picked up and, at that point, the yetis are inactive because they
haven't even been activated yet.

Fixed by making sure Lara is in yeti jail when the level flips, making
the mask flip ineligible for the achievement.