[TR2][33517] avoid corner bug detection when climbing small ledges
When Lara climbs a small ledge, her vertical position is not updated
smoothly. Instead, at the end of the animation, it is suddenly updated
from the height of the lower floor to the height of the upper floor.
This is supposed to be protected against by checking Lara's current
action: since she's not running, swimming or upwards jumping, the
protection shouldn't be triggered by the sudden vertical position
update.

However, if the player starts running right after the climb (by holding
a direction during the climb), it's possible for the game to update
Lara's vertical position on the same frame as it updates her action to
running, causing the corner bug protection to trigger.

To fix this problem, we now also look at Lara's action on the previous
frame, and it must also be either running, swimming or upwards jumping.
This avoids the mistrigger post-ledge-climb, but still triggers on legit
corner bugs, as those always happen mid-action.