[TR2][33777] use level-ready as starting points instead of level-loaded
Whole-level challenges used the is-loading flag turning off as the
starting point. In e2158698ec4223dec54332cc56e852146a7758b8, it was
found out that there's one frame between this flag turning off and the
level being ready (entities initialized, etc...). As a result, during
that one frame, it was possible for stale data to be used and cause
the achievement to misbehave.

As a specific example, if Tomb Hiker was started right after booting up
the game, Lara's vehicle-id field would have the value 0 during that
frame, which is interpreted as Lara being in a vehicle, causing the
achievement to reset right away. If Lara had already been in another
level, the vehicle-id field would most likely contain -1, and the issue
would not manifest. If Lara loaded the level while in a vehicle, the
issue would still manifest, of course.

Fixed in the same manner that the original corner bug issue was fixed:
by using the level-ready frame as the starting point. This frame is
detected through the music track id changing from zero to non-zero.