[sm74ee][33811] count blue coins using collision with mario
In 4e63d733c9a03480384a1e89da03bdfc0ddfe432 the blue coin achievements
were changed to count coins as having been collected when they unloaded
on a frame in which mario got at least 5 coins. As noted then, the
solution was not perfect, but I miscalculated how likely a mistrigger
would be. If the player collected the 2nd to last coin on the same frame
that the coins despawn, the achievement would trigger, because the last
coin would also unload on a frame in which 5 coins had been collected.

The achievements have now been rewritten for the 3rd time, hopefully the
last. Blue coins are now considered collected if they unload on a frame
in which they're colliding with Mario. This _should_ be a foolproof
solution....