[TR2][33777] detect level-ready/playable using a new flag
A previous attempt at a fix used the music track id as the way to detect
when a levels becomes ready / playable. This turned out to not work,
because there are some levels that start without any track playing.

The level-ready event is now detected using a new flag that seems to
only be set during normal gameplay. It's unset during loading,
cutscenes, post-level, etc...