[TR2][35062] fix swan dive detection at the end of xian's slide
The swan dive used to be detected by checking for the start of the swan
dive action whilst Lara was in one of the 2 rooms that make up the end
of the slide: 188 and 189. However, if the swan dive is started right
at the edge of the slide, the swan dive action only starts when Lara is
already in the next room: 8.

Fixed by allowing the swan dive to start at room 8. In addition,
because room 8 is large enough to include locations that are not at the
edge of the slide (including the gold secret alcove), the achievement
now also restricts Lara's coordinates.