[jptlw] avoid trigger when stegosaurus carcass explodes off-screen
There's a bug in the game that sometimes causes the stegosaurus carcass
in Activate Sensors to explode right after spawning without player
intervention. This change avoids triggering the achievement if that
happens.

The downside is that the achievement will be impossible on some visits
to Activate Sensors, but the bug seems rare enough for this to not
result in a huge inconvenience to the player.