[jptlw] workaround ingame_timer incompatibility between cores
The lazor achievement was checking whether the player's last-shot-time
was equal to the value of the ingame timer in order to determine if the
player had just shot the taser. This worked fine on Genesis Plus GX.

In picodrive, the last-shot-time seems to lag by one in comparison to
the ingame timer for unknown reasons, probably related to the timer's
emulation. This cause the achievement to not trigger.

Fixed by also allowing the last-shot-time to be equal to the ingame
timer minus one. This works on both Genesis Plus GX and picodrive.