[dino crisis][37360] account for extra gail ending variation
When the player chooses Gail's idea in the final decision, there are
actually two rooms in which Gail can be in: the Stabilizer Experiment
Room, which was already accounted for, and Special Weapons Storage. It's
not clear at the moment how the location is chosen, but it's probably
related to the choice made in previous decision.