[trc] remove loadable level start points
Since the release of the set, it has become apparent that the system I
implemented to allow saving at the start of levels to retry achievements
and leaderboards is both causing problems and unintuitive.

The problem is that it disallows the player from moving right away,
instead forcing the player to wait until the leaderboard starts to
actually move. This is not necessarily a clear requirement even when
leaderboards are active, and many people don't have them enabled at all.
Worst of all is that breaking this unclear rule nullifies NLNMAS
attempts, the hardest achievements of the set. In some levels it's
possible to have both retryable saves at the start and allow the player
to move right away, but there are quite a few where it's not possible:
colosseum, the base, deepsea dive, etc...

As such, I have decided to remove this system. Players will have to save
at the end of the previous level or use the skip codes in the title
screen. While annoying due to cutscenes, I think it'll be a much more
intuitive approach that anyone will understand, even if they don't know
the details of the set or don't even use the site, for example.