[chimera] don't trigger torch warhead achievement with pyramid warhead
Even through the torch warhead must be the third one used in order to
beat the game, it's possible to use the pyramid warhead as the third
one, resulting in an unbeatable puzzle due to the torch being replaced
with a mug.

The warhead achievements were made expecting the torch warhead to be
the third, and the pyramid warhead the fourth. This assumption is wrong
as explained.

Fixed in two ways:

* The achievements no longer refer to the warheads as the first, second,
etc... Instead, the warheads are identified solely by their components.

* The triggers for the torch and pyramid warhead achievements were
modified to account for the two possible orders. The torch achievement
only triggers if the warhead being activated is the 3rd one and the
torch has been collected. The pyramid achievement triggers if the
warhead is the 4th one, or if it's the 3rd one and the torch has not
been collected.