[soul reaver][65290] avoid issues with event vars on room transitions
The rahab-fast achievement trigger relies on event flags switching when
the boss fight starts and ends. These flags are down a pointer chain and
the final pointers change whenever the room raziel is in changes. The
issue here was that when coming back to the room after having defeated
the boss, the delta check on the pointer chain would be looking at
values from the previous room. It so happens that they matched the
needed values often, though not always.

Fixed by checking whether the final pointer to the event flags is the
same in the previous and current frame. This way it's no longer possible
for the values from the previous room to get used.

Note: RATools v1.61.1 can't compile the achievement's original AddHits
chain. A modified RATools was used to replicate the original logic. Also
had to flip the order of the or'd conditions for the same reason.