[soul reaver][96068] avoid possible first-frame event flag issue
The "Complacent in Their Arrogance" achievement, like most reach-area
achievements, uses a event flag that signals whether the area has been
entered before. The event flags are behind a pointer chain that only
points to our target when we enter the area. As such, we need at least
one frame after entering, in which the flag is not set. If the flag gets
set on the very first frame, somehow, the achievement won't trigger.
There is precedent for this: "Through the Fallen Skull" does not use
event flags because the flag always gets set on the first frame. Perhaps
it can also happen for this achievement, but only rarely...

As such, I have reimplemented the achievement using the same approach as
"Through the Fallen Skull": trigger if raziel gets put into a cutscene
during the first second of entering the area.