[soul reaver][67981] better detection of sunlight deaths
This achievement was quite noticeably broken, just from the fact that it
gets triggered by way too many players, way too early.

The trigger looks for a vampire instance that has just started dying
from fire. Fire, however, can be triggered in many ways, not just
sunlight. As such, the trigger excluded a bunch of scenarios based on
the prior state of the vampire instance. The exclusion list was not
enough though, and I managed to replicate the issue in multiple ways,
namely using a torch on a vampire that hasn't noticed Raziel or using
the fire and sunlight glyphs.

I have found a new damageType field of monster instances that can be
used to differentiate between deaths from actual fire and sunlight. The
sunlight value only gets used by beams of light or the sunlight glyph.
The latter is excluded using Raziel's animation value. In addition,
many of the prior-state-based exclusions remain, because we want the
vampire to walk into the sunlight themselves, not get pushed into it by
Raziel.