[pokemon stadium][93774] reimplement mewtwo hall-of-fame achievement
The achievement was implemented using hall-of-fame data. This data lives
in SRAM and is only temporarily loaded into RAM when it needs to be read
or modified. The trigger needed the hall-of-fame data to be in memory
for one frame with mewtwo unset, but it seems it's possible for mewtwo
to already be set on the very first frame.

Fixed by reimplementing the achievement without using hall-of-fame data.
It now triggers when mewtwo is shown during the victory lap after
beating a cup or the rival in gym leader castle.