[tr2][67789] avoid edge case mistrigger of ice queen achievement
The achievement uses the moment the second round of snow stops moving to
trigger. However, if Lara is at the very bottom of the hill as described
in the report, the snow stops a couple frames before Lara dies, causing
the mistrigger.

Fixed by delaying the achievement trigger for 15 frames. The original
trigger is stored in a hit, and 15 hits are then required with the snow
stopped and at x position >= 81000.