[tr2][98355] properly detect whether venice mines have blown up
The venice-mines achievement used a mine's active flag to determine
whether the mines were still there. It seems, however, that the mines
remain active even after being destroyed, which is a massive
oversight...

Fixed by checking whether the mine is collidable. This flag does change
when the mines get destroyed.