[tr2][98406] consider on-boat finish in venice mines achievement
The achievement is triggered when the level is beat. The
challenge-started hit is reset when the final doors fully close and
lara is not in the past-the-doors room.

One issue was that there are actually two past-the-doors rooms, one for
the water and one for above-water. When beating the level on a boat,
lara would be in the above-water room, which the achievement did not
consider as being past-the-doors.

Fixed by considering the above-water past-the-doors room in the reset
logic.

In addition, a delay was introduced to the reset, providing some leeway
for possible issues that although not observed, might be possible. One
such possibility would be for the doors to fully close while lara was
still not in the past-the-doors rooms, but moving with enough speed to
"clip" into them right after.