[tr2][77514] fix challenge start in venice mines achievement
The achievements stored a hit when the door timer was running and the
door animation changed from closed to opening. This works fine under
normal conditions, because the door opens during the cutscene, closes,
and then opens again when the timer starts.

However, a player took it to the next level by canceling the cutscene,
which I didn't know was even possible, and then jumping to the boat
positioned right on the ramp. This sped up version of play meant the
double opening of the door didn't happen, and thus the challenge was
never considered to have started. Too good for their own good, one might
say.

Fixed by storing the start hit when the door's animation changes to
open. This seems to work for both situations.