[tr2] remove corner bug protection from the set
The corner bug protection works ok, but has issues. On the one hand, it
can't detect every instance of the bug or other similar bugs. On the
other, it can mistrigger, leading to user confusion and tickets. It
seems mistriggers are very rare, but some tickets seem to stem from
that.

In addition, the protection only exists in Tomb Raider 2's set, even
though the bug exists in all the other games, which is inconsistent.

Removed the protection. More trouble than its worth...